// src/app/models/user.ts
export class User {

    id: number;

    name: string;

    email: string;

    phone: string;

    plan_id: number;

    company: string;

    company_country: string;

    company_city: string;

    company_zipcode: string;

    company_address: string;

    password: string;
    
    reset_code: string; 

}